AP_AnalogSource: make read_average() return a float
this allows it to take advantage of better than 1 bit resolution by
averaging

+++ libraries/AP_AnalogSource/AP_AnalogSource_Arduino.h
@@ -32,7 +32,7 @@ class AP_AnalogSource_Arduino : public AP_AnalogSource
 
     // read the average 16 bit ADC value since
     // we last called read_average().
-    uint16_t read_average(void);
+    float read_average(void);
 
     private:
     uint8_t _pin_index;
